Clarified arc fitting (#27)
* Update arc fitting feature description and notes

Clarified the arc fitting feature's impact on print quality and G-code size. Added notes on CPU usage and recommendations for Klipper machines.
The idea is to stop with the myth that arcs print better than segments, since usually it's not the case

* Clarify arc fitting feature description

formatting

* Update quality_settings_precision.md for clarity

Clarify the impact of arc fitting on print quality and CPU usage.

* Correct segment length from 0.6 mm to 0.4 mm

Updated the segment length comparison in quality settings documentation.

* Refine segment length details in quality settings

Updated the segment length range for OnShape's default preset 'Fine' in the quality settings documentation.

* Refine arc fitting feature explanation in documentation

Shortened

* Update quality_settings_precision.md

* Improve clarity on arc fitting and its effects

Clarified the explanation of arc fitting and its implications for print quality, especially for low-quality STLs and older printers. Added notes on CPU usage and provided a link for Klipper printer configuration.

* Refine documentation on arc fitting feature

Clarified the purpose and implications of arc fitting in G-Code generation. Improved readability and precision of the documentation regarding print quality and printer performance.

* Clarify usage of arc resolution settings in documentation

Added a note about the use of arc resolution settings.

* Update print_settings/quality/quality_settings_precision.md

Co-authored-by: Ian Bassi <ian.bassi@outlook.com>

* Revise arc fitting section for clarity and accuracy

Clarified the purpose and implications of arc fitting in G-Code generation, emphasizing the need for printer compatibility and potential quality impacts.

---------

Co-authored-by: Ian Bassi <ian.bassi@outlook.com>